pyproject.toml
[tool.poetry]
name = "flaskgpt"
version = "0.1.0"
description = "ChatGPT clone using Flask"
authors = ["Arthur Bussiere <arthur.bussiere@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "2.3.3"
openai = "0.27.10"
python-dotenv = "1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
